requirements.txt
fastapi
pydantic
uvicorn
pymongo
python-jose
passlib[bcrypt]
python-dotenv